[issue-6] FractalParser is a separate thing from Fractal
Each existing fractal gets its own separate parser.
