+++ apps/fractals/lib/fractals/escape_time/burning_ship_parser.ex
@@ -1,0 +1,15 @@
+defmodule Fractals.EscapeTime.BurningShipParser do
+  @moduledoc """
+  Parses a config for the burning ship fractal.
+  """
+
+  @behaviour Fractals.FractalParser
+
+  @impl Fractals.FractalParser
+  def parse(_params) do
+    %Fractals.Fractal{
+      type: :burning_ship,
+      module: Fractals.EscapeTime.BurningShip
+    }
+  end
+end
