+++ apps/fractals/lib/fractals/escape_time/julia_parser.ex
@@ -1,0 +1,16 @@
+defmodule Fractals.EscapeTime.JuliaParser do
+  @moduledoc """
+  Parses a config for a Julia fractal.
+  """
+
+  @behaviour Fractals.FractalParser
+
+  @impl Fractals.FractalParser
+  def parse(params) do
+    %Fractals.Fractal{
+      type: :julia,
+      module: Fractals.EscapeTime.Julia,
+      algorithm_params: %{c: Complex.parse(params[:c])}
+    }
+  end
+end
